+++ cli/global.d.ts
@@ -1,3 +1,1 @@
-declare module "pem-file";
 declare module "get-folder-size";
-declare module "decomp-tarxz";